HELP!!!!!!
The struggle is real. Hello all. I am a newby to Python and I have created a program that will stop, start, or restart a windows service. I am using it as my learning tool for the language. I have simplified the code, now I am trying to Create classes or a class and use the init method and add decorators. Can any help?